Tweak vim config to work without plugins installed
This config won't throw any errors if a plugin is missing. There are
still keybdings created which could fail, but that's fine.  I want to
see an error if I try to do something that isn't installed, but I don't
want to get an error just for opening up a fresh vim install