Add katago packages
This set of packages is currently what's needed to build katago.
Since I don't have a modern GPU I have to build it using the "Eigen"
option which uses the CPU, but is surprisingly effective!

Katrain is a whole other level on top, and probably when I next come to
do it I'll have to figure it out again, because I don't think it's in a
stable state.