Clean up all the xmobar scripts
The biggest change here is that I've made the mail checking for
gargantua-two async. The result of that is that there are now no network
operations in the execution flow of xmobar and the volume meter is
responsive!

I've also cleaned up bunch of the scripts to eliminate hardcoded
variables and improve readability.